+++ cmd/import_test.go
@@ -15,10 +15,14 @@
 package cmd
 
 import (
-	"cloud.google.com/go/spanner"
 	"context"
 	"flag"
 	"fmt"
+	"strings"
+	"testing"
+	"time"
+
+	"cloud.google.com/go/spanner"
 	spannerclient "github.com/GoogleCloudPlatform/spanner-migration-tool/accessors/clients/spanner/client"
 	spanneraccessor "github.com/GoogleCloudPlatform/spanner-migration-tool/accessors/spanner"
 	"github.com/GoogleCloudPlatform/spanner-migration-tool/common/constants"
@@ -27,9 +31,6 @@ import (
 	"github.com/GoogleCloudPlatform/spanner-migration-tool/spanner/ddl"
 	"github.com/google/subcommands"
 	"github.com/stretchr/testify/assert"
-	"strings"
-	"testing"
-	"time"
 )
 
 const expectedDDL = "CREATE TABLE cart ( \tuser_id STRING(20) NOT NULL , \tproduct_id STRING(20) NOT NULL , \tquantity INT64, \tlast_modified TIMESTAMP NOT NULL , ) PRIMARY KEY (user_id, product_id);CREATE INDEX idx ON cart (quantity)"
@@ -50,6 +51,122 @@ func TestBasicCsvImport(t *testing.T) {
 	importDataCmd.csvLineDelimiter = "\n"
 	importDataCmd.csvFieldDelimiter = ","
 	importDataCmd.Execute(context.Background(), fs)
+}
+
+func TestImportDataCmd_SetFlags(t *testing.T) {
+	cmd := &ImportDataCmd{}
+	fs := flag.NewFlagSet("import", flag.ContinueOnError)
+	cmd.SetFlags(fs)
+
+	assert.NotNil(t, fs.Lookup("instance-id"))
+	assert.NotNil(t, fs.Lookup("database-name"))
+	assert.NotNil(t, fs.Lookup("table-name"))
+	assert.NotNil(t, fs.Lookup("source-uri"))
+	assert.NotNil(t, fs.Lookup("source-format"))
+	assert.NotNil(t, fs.Lookup("schema-uri"))
+	assert.NotNil(t, fs.Lookup("csv-line-delimiter"))
+	assert.NotNil(t, fs.Lookup("csv-field-delimiter"))
+	assert.NotNil(t, fs.Lookup("project"))
+}
+
+func TestValidateInputLocal_MissingInstanceID(t *testing.T) {
+	input := &ImportDataCmd{}
+	err := validateInputLocal(input)
+	assert.Error(t, err)
+	assert.Contains(t, err.Error(), "Please specify instanceId")
+}
+
+func TestValidateInputLocal_MissingDatabaseName(t *testing.T) {
+	input := &ImportDataCmd{instanceId: "test-instance"}
+	err := validateInputLocal(input)
+	assert.Error(t, err)
+	assert.Contains(t, err.Error(), "Please specify databaseName")
+}
+
+func TestValidateInputLocal_MissingSourceURI(t *testing.T) {
+	input := &ImportDataCmd{instanceId: "test-instance", databaseName: "test-db"}
+	err := validateInputLocal(input)
+	assert.Error(t, err)
+	assert.Contains(t, err.Error(), "Please specify sourceUri")
+}
+
+func TestValidateInputLocal_MissingSourceFormat(t *testing.T) {
+	input := &ImportDataCmd{instanceId: "test-instance", databaseName: "test-db", sourceUri: "file:///tmp/data.csv"}
+	err := validateInputLocal(input)
+	assert.Error(t, err)
+	assert.Contains(t, err.Error(), "Please specify sourceFormat")
+}
+
+func TestValidateInputLocal_CSVMissingSchemaURI(t *testing.T) {
+	input := &ImportDataCmd{instanceId: "test-instance", databaseName: "test-db", sourceUri: "file:///tmp/data.csv", sourceFormat: constants.CSV}
+	err := validateInputLocal(input)
+	assert.Error(t, err)
+	assert.Contains(t, err.Error(), "Please specify schemaUri")
+}
+
+func TestValidateInputLocal_SuccessCSV(t *testing.T) {
+	input := &ImportDataCmd{
+		instanceId:   "test-instance",
+		databaseName: "test-db",
+		sourceUri:    "file:///tmp/data.csv",
+		sourceFormat: constants.CSV,
+		schemaUri:    "file:///tmp/schema.csv",
+	}
+	err := validateInputLocal(input)
+	assert.NoError(t, err)
+}
+
+func TestValidateInputLocal_SuccessNonCSV(t *testing.T) {
+	input := &ImportDataCmd{
+		instanceId:   "test-instance",
+		databaseName: "test-db",
+		sourceUri:    "gs://bucket/data.avro",
+		sourceFormat: "avro",
+	}
+	err := validateInputLocal(input)
+	assert.NoError(t, err)
+}
+
+func TestHandleTableNameDefaults_TableNamePresent(t *testing.T) {
+	tableName := "explicit_table"
+	sourceUri := "gs://bucket/data.csv"
+	result := handleTableNameDefaults(tableName, sourceUri)
+	assert.Equal(t, "explicit_table", result)
+}
+
+func TestHandleTableNameDefaults_TableNameEmptyFileScheme(t *testing.T) {
+	tableName := ""
+	sourceUri := "file:///path/to/my_data.csv"
+	result := handleTableNameDefaults(tableName, sourceUri)
+	assert.Equal(t, "my_data.csv", result)
+}
+
+func TestHandleTableNameDefaults_TableNameEmptyGCScheme(t *testing.T) {
+	tableName := ""
+	sourceUri := "gs://my-bucket/data_file.txt"
+	result := handleTableNameDefaults(tableName, sourceUri)
+	assert.Equal(t, "data_file.txt", result)
+}
+
+func TestHandleTableNameDefaults_TableNameEmptyLocalPathNoScheme(t *testing.T) {
+	tableName := ""
+	sourceUri := "/tmp/another_file.json"
+	result := handleTableNameDefaults(tableName, sourceUri)
+	assert.Equal(t, "another_file.json", result)
+}
+
+func TestHandleTableNameDefaults_TableNameEmptyRelativePath(t *testing.T) {
+	tableName := ""
+	sourceUri := "relative/path/some_data.avro"
+	result := handleTableNameDefaults(tableName, sourceUri)
+	assert.Equal(t, "some_data.avro", result)
+}
+
+func TestHandleTableNameDefaults_TableNameEmptyURIWithTrailingSlash(t *testing.T) {
+	tableName := ""
+	sourceUri := "gs://my-bucket/folder/"
+	result := handleTableNameDefaults(tableName, sourceUri)
+	assert.Equal(t, "folder", result)
 }
 
 func TestImportDataCmd_HandleDumpExecute(t *testing.T) {
